ref(replay): Extract integration to clarify public API
